Only include protobuf wrapper for PyTorch
