+++ CMakeLists.txt
@@ -296,7 +296,9 @@ include_directories(BEFORE ${PROJECT_SOURCE_DIR})
 include_directories(BEFORE ${PROJECT_BINARY_DIR})
 
 # ---[ Old caffe protobuf
-add_subdirectory(caffe/proto)
+if(BUILD_CAFFE2)
+  add_subdirectory(caffe/proto)
+endif()
 
 # ---[ Main build
 add_subdirectory(caffe2)
