+++ caffe2/utils/CMakeLists.txt
@@ -1,15 +1,24 @@
-set(Caffe2_CPU_SRCS ${Caffe2_CPU_SRCS}
-        utils/proto_utils.cc
-        utils/murmur_hash3.cc
-        utils/smart_tensor_printer.cc
-        utils/signal_handler.cc
-        utils/string_utils.cc
-        utils/threadpool/ThreadPool.cc
-        utils/cpuid.cc
-        utils/bench_utils.cc
-        utils/math_cpu.cc
-        utils/math_utils.cc
-        )
+list(APPEND Caffe2_CPU_SRCS
+  utils/proto_wrap.cc)
+
+# ---[ only support the above when full caffe2 isn't built
+if (NOT BUILD_CAFFE2)
+  set(Caffe2_CPU_SRCS ${Caffe2_CPU_SRCS} PARENT_SCOPE)
+  set(Caffe2_GPU_SRCS ${Caffe2_GPU_SRCS} PARENT_SCOPE)
+  return()
+endif()
+
+list(APPEND Caffe2_CPU_SRCS
+  utils/proto_utils.cc
+  utils/murmur_hash3.cc
+  utils/smart_tensor_printer.cc
+  utils/signal_handler.cc
+  utils/string_utils.cc
+  utils/threadpool/ThreadPool.cc
+  utils/cpuid.cc
+  utils/bench_utils.cc
+  utils/math_cpu.cc
+  utils/math_utils.cc)
 
 # ---[ threadpool/pthreadpool* is a local modification of the NNPACK
 # pthreadpool with a very similar interface. Neither NNPACK, nor this
